New minimum of post iris pegin.
Functionalities are added to obtain this new minimum. The way a transaction is considered to be a pegin type has been changed. Pegin is considered if it is of this type, separating the validation from the minimum in other functionality. New functionality is created that unites the verification of the type of transaction together with the validation of the minimum according to RSKIP.
It also requires changes in isMigrationTx, but only to be able to include activation as a parameter, to be used internally by isPeginTx.